+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Add ability to start game from world selection screen. Can be started in combined and dedicated server mode.
 
 ### Changed
+- Start server internally in minecraft import tool.
 - Move all engine initialization code into enginestarter.d.
 - Create PluginManager outside of client and server plugins.
 - Redo netlib package.